Migrate to Spring Boot 2.2 + Kotlin.
While upgrading to Spring Boot 2.2, this project presented a unique opportunity to start from scratch, and code it all in Kotlin. This proved as a handy way for me to not only warm up my Kotlin programming talents, but to also refashion this application to take advantage of key features.

Also able to drop stuff that was no longer necessary, like Spring Social Github, a project that never reached GA. Instead, it turns out that WebClient easily supports applying a header in order to talk to github, and it's also super easy to create the DTOs needed to extract github information.

The sacrifice was letting go of inuitcss. That project, which is still in alpha, presented a clean, lightweight CSS toolkit to build a responsive UI. Unfortunately, they overhauled the whole thing since last used, and some of the key components were removed. For now, I've removed it from the build process. I've retained the color palette, but the look and feel of this app will have to be revisited to restore it.